requirements.txt
streamlit
librosa
numpy
streamlit-webrtc
google-generativeai
python-dotenv
assemblyai
fastapi
uvicorn
jinja2


